DELIA-39530: Wrong position calculated in isobmffprocessor
Reason for change: Fix position and pts calculation rounding off
due to typecasting. Enable some crucial logs by default
Test Procedure: tested on pace xi5
Risks: Low

+++ isobmff/isobmffbox.cpp
@@ -471,7 +471,7 @@ TfdtBox* TfdtBox::constructTfdtBox(uint32_t sz, uint8_t *ptr)
 	}
 	else
 	{
-		mdt = READ_U32(ptr);
+		mdt = (uint32_t)READ_U32(ptr);
 	}
 	FullBox fbox(sz, Box::TFDT, version, flags);
 	return new TfdtBox(fbox, mdt);
